dev-requirements.txt
# SPDX-FileCopyrightText: 2024 PNED G.I.E.
#
# SPDX-License-Identifier: Apache-2.0

black==24.8.0
pytest==8.3.2
requests-mock==1.12.1
pytest-cov==5.0.0